Significantly redesign Media Tower. Fix bolter shooting through enemy force fields. Fix parkour infinite wall-run exploit. Fix parkour mantle clipping through thin ceilings. Limit max purchase quantity. Tweak audio. Fix parkour wall-run footstep particles appearing delayed.

+++ src/game/game.h
@@ -105,7 +105,7 @@ struct Game
 		s32 locke_index;
 		ZoneState zones[MAX_ZONES];
 		Group group;
-		s16 resources[s32(Resource::count)];
+		s32 resources[s32(Resource::count)];
 		AssetID zone_last;
 		AssetID zone_current;
 		AssetID zone_overworld;
